moved and fixed test of _hadoop_conf_args()
